move duration converter to be not overtime specific
