+++ src/main/java/org/synyx/urlaubsverwaltung/overtime/Overtime.java
@@ -1,5 +1,6 @@
 package org.synyx.urlaubsverwaltung.overtime;
 
+import org.synyx.urlaubsverwaltung.DurationConverter;
 import org.synyx.urlaubsverwaltung.person.Person;
 
 import javax.persistence.Column;
