+++ src/test/java/org/synyx/urlaubsverwaltung/overtime/DurationConverterTest.java
@@ -2,6 +2,7 @@ package org.synyx.urlaubsverwaltung.overtime;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.synyx.urlaubsverwaltung.DurationConverter;
 
 import java.time.Duration;
 
@@ -18,10 +19,16 @@ class DurationConverterTest {
 
     @Test
     void ensureConversionOfMinutesCorrectly() {
-        for (int i = 0; i <= 60; i++) {
+        for (int i = 0; i <= 61; i++) {
             Duration expected = Duration.ofMinutes(i);
             final Duration actual = sut.convertToEntityAttribute(sut.convertToDatabaseColumn(expected));
             assertThat(actual).isEqualTo(expected);
         }
+    }
+
+    @Test
+    void ensureNullReturnsNull() {
+        assertThat(sut.convertToEntityAttribute(null)).isNull();
+        assertThat(sut.convertToDatabaseColumn(null)).isNull();
     }
 }
